Re-implement PersistentMessageQueue with support for multiple messages under the same key

+++ Subverse.Server/PersistientMessageQueue.cs
@@ -17,12 +17,15 @@ namespace Subverse.Server
             _db = new LiteDatabase(_configuration.GetConnectionString("MessageQueueImpl"));
         }
 
-        public Task<KeyedMessage> DequeueAsync()
+        public Task<KeyedMessage?> DequeueAsync()
         {
             var collection = _db.GetCollection<KeyedMessage>();
-            var keyedMessage = collection.Query().FirstOrDefault();
+            var keyedMessage = collection.FindAll().First();
 
-            collection.Delete(keyedMessage.Id);
+            if (keyedMessage is not null)
+            {
+                collection.Delete(keyedMessage.Id);
+            }
 
             return Task.FromResult(keyedMessage);
         }
@@ -30,17 +33,20 @@ namespace Subverse.Server
         public Task<SubverseMessage?> DequeueByKeyAsync(string key)
         {
             var collection = _db.GetCollection<KeyedMessage>();
-            var keyedMessage = collection.FindById(key);
+            var keyedMessage = collection.FindOne(x => x.Key == key);
 
-            collection.Delete(keyedMessage.Id);
+            if (keyedMessage is not null)
+            {
+                collection.Delete(keyedMessage.Id);
+            }
 
-            return Task.FromResult<SubverseMessage?>(keyedMessage.Message);
+            return Task.FromResult(keyedMessage?.Message);
         }
 
         public Task EnqueueAsync(string key, SubverseMessage message)
         {
             var collection = _db.GetCollection<KeyedMessage>();
-            var keyedMessage = new KeyedMessage(key, message);
+            var keyedMessage = new KeyedMessage(collection.LongCount(), key, message);
 
             collection.Insert(keyedMessage);
